requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
-e git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1#egg=clip
cloudpathlib==0.21.1
cloudpickle==3.1.0
cmake==3.31.4
colorama==0.4.6
comm==0.2.2
compressai==1.1.5
conda-pack==0.8.1
confection==0.1.5
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
drjit==0.4.6
einops==0.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.14.0
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.9.0
ftfy==6.2.0
future==1.0.0
gast==0.4.0
gem-torch==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
horovod==0.28.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
intel-cmplr-lib-ur==2025.0.4
intel-openmp==2025.0.4
ipydatawidgets==4.3.2
ipykernel==6.29.5
ipython==8.31.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-widgets==3.0.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
lit==18.1.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mitsuba==3.5.2
mkl==2025.0.1
ml-dtypes==0.2.0
mmcls==0.25.0
mmcv==2.2.0
mmcv-full==1.6.2
mmengine==0.8.4
mmsegmentation==0.27.0
model-index==0.1.11
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.13
myutils==0.0.21
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==6.0.0
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.30.0
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.10
prettytable==3.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pythreejs==2.4.2
pytorch-msssim==1.0.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.3
scikit-image==0.25.0
scikit-learn==1.7.0
scipy==1.14.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
sionna==0.19.1
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-cu12==10.7.0
tensorrt-cu12-bindings==10.7.0
tensorrt-cu12-libs==10.7.0
tensorrt-libs==8.6.1
termcolor==2.5.0
terminado==0.18.1
thinc==8.2.5
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.12.12
timm==1.0.17
tinycss2==1.4.0
tokenizers==0.15.2
tomli==2.2.1
torch @ file:///media/bohnsix/ProxyCLIP/torch-1.13.1%2Bcu117-cp310-cp310-linux_x86_64.whl#sha256=14c5c9db09df8cf1b3942a3479c779da6e293a84a162d8a6ac71e2bde30e30c5
torch-geometric==2.6.1
torchaudio==0.13.1
torchmetrics==1.5.2
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
traittypes==0.2.1
transformers==4.37.2
triton==2.0.0
ttach==0.0.3
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
umf==0.9.1
uri-template==1.3.0
urllib3==1.26.20
wandb==0.19.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.14.1
yapf==0.40.1
yarl==1.18.3
zipp==3.21.0
